#956 Phase 8: extract cos/cross_binding.rs from tx.rs (FINAL phase)
Per docs/pr/956-phase8-cross-binding/plan.md (v2). The final
phase of #956 — closes out the cos/ submodule decomposition.

Move list (~250 LOC):
- 2 types: Step1Action enum, LocalRoutingDecision struct
- 8 fns: resolve_local_routing_decision, cos_fast_interface,
  cos_fast_queue, redirect_local_cos_request_to_owner +
  _binding (test-only), prepared_cos_request_stays_on_current_tx_binding,
  redirect_prepared_cos_request_to_owner + _binding

Visibility:
- pub(in crate::afxdp): all 7 production-callable items
- cfg-gated: redirect_local_cos_request_to_owner_binding (test-only)

#[inline] preserved/added on per-byte hot path. Larger memcpy-style
prepared variants follow Phase 7 precedent (un-attributed).

One back-edge: tx::recycle_prepared_immediately (visibility bumped
to pub(in crate::afxdp)). Stays in tx.rs because it touches XSK
ring frame recycling — worker-binding territory.

Visibility bumps in sibling files: umem::BindingLiveState +
types::WorkerCommand → pub(in crate::afxdp).

Net change: -249 lines from tx.rs.

Tests: cargo test --bins -> 865 passed, 0 failed, 2 ignored.

Reviewers: Codex round-1 caught scope incompleteness (added
Step1Action, LocalRoutingDecision, 3 helpers); v2 expanded.
Gemini round-1 reached the same architectural conclusions
independently.

After this phase the cos/ submodule decomposition (#956) is
COMPLETE. tx.rs left behind is XSK ring management + worker-
binding glue + the deferred TX-completion family (apply_cos_*_result,
restore_*_inner, transmit_*) which Phase 7 bumped to
pub(in crate::afxdp). Future TX-completion / worker-binding
extractions can proceed against the cleaner tx.rs surface.
