+++ userspace-dp/src/afxdp/tx.rs
@@ -1124,255 +1124,6 @@ fn ingest_cos_pending_tx_with_provenance(
     bound_pending_tx_local(binding);
 }
 
-/// #780: Step 1 action variants. Mirrors the action taken inside
-/// `redirect_local_cos_request_to_owner` after the bail checks
-/// have been passed.
-#[derive(Clone)]
-enum Step1Action {
-    /// The owner worker's owner_live arc is directly addressable
-    /// (fast path).
-    Arc(Arc<BindingLiveState>),
-    /// Fall back to the per-worker command channel (slow path).
-    Command(u32),
-}
-
-/// #780: routing-decision cache value. Carries BOTH Step 1 and
-/// Step 2 options so the dispatch in `ingest_cos_pending_tx_with_provenance`
-/// can fall through Step 1 → Step 2 → Step 3 (EnqueueLocal) on
-/// Err at each boundary — exact cascade semantics of the
-/// pre-#780 three-function chain. Codex review round 2 flagged
-/// the previous revision's lack of fallthrough as a HIGH
-/// semantic regression.
-#[derive(Clone)]
-struct LocalRoutingDecision {
-    /// `None` when Step 1 bails (queue absent, shared_exact-with-
-    /// owner, or owner_worker_id == current_worker_id). Present
-    /// when Step 1 would route.
-    step1: Option<Step1Action>,
-    /// `None` when Step 2 bails (iface absent, no tx_owner_live,
-    /// or ptr_eq(tx_owner_live, current_live)). Present when
-    /// Step 2 would route.
-    step2: Option<Arc<BindingLiveState>>,
-}
-
-/// #780: resolve the routing decision for a (iface, queue) pair.
-/// Preserves the exact pre-#780 cascade semantics. Moved out of
-/// the closure so it can be unit-tested independently. Carries
-/// BOTH step options in the returned decision so dispatch can
-/// walk the same fallthrough as the original cascade when an
-/// earlier step's enqueue returns Err.
-fn resolve_local_routing_decision(
-    iface_fast_opt: Option<&WorkerCoSInterfaceFastPath>,
-    cos_queue_id: Option<u8>,
-    current_worker_id: u32,
-    current_live: &Arc<BindingLiveState>,
-) -> LocalRoutingDecision {
-    let mut step1: Option<Step1Action> = None;
-    let mut step2: Option<Arc<BindingLiveState>> = None;
-    if let Some(iface_fast) = iface_fast_opt {
-        // Step 1 (mirrors redirect_local_cos_request_to_owner):
-        if let Some(queue_fast) = iface_fast.queue_fast_path(cos_queue_id) {
-            let step1_bail = (queue_fast.shared_exact && iface_fast.tx_owner_live.is_some())
-                || queue_fast.owner_worker_id == current_worker_id;
-            if !step1_bail {
-                step1 = Some(match queue_fast.owner_live.as_ref() {
-                    Some(arc) => Step1Action::Arc(arc.clone()),
-                    None => Step1Action::Command(queue_fast.owner_worker_id),
-                });
-            }
-        }
-        // Step 2 (mirrors redirect_local_cos_request_to_owner_binding):
-        // ALWAYS evaluated — the old cascade ran Step 2 after Step 1
-        // returned Err, so Step 2 is reachable whether or not Step 1
-        // also routes. We cache both here; the dispatch loop walks
-        // Step 1 first, falling through to Step 2 on Err.
-        if let Some(owner_live) = iface_fast.tx_owner_live.as_ref() {
-            if !Arc::ptr_eq(owner_live, current_live) {
-                step2 = Some(owner_live.clone());
-            }
-        }
-    }
-    LocalRoutingDecision { step1, step2 }
-}
-
-#[inline]
-fn cos_fast_interface<'a>(
-    cos_fast_interfaces: &'a FastMap<i32, WorkerCoSInterfaceFastPath>,
-    egress_ifindex: i32,
-) -> Option<&'a WorkerCoSInterfaceFastPath> {
-    cos_fast_interfaces.get(&egress_ifindex)
-}
-
-#[inline]
-fn cos_fast_queue<'a>(
-    cos_fast_interfaces: &'a FastMap<i32, WorkerCoSInterfaceFastPath>,
-    egress_ifindex: i32,
-    requested_queue_id: Option<u8>,
-) -> Option<(&'a WorkerCoSInterfaceFastPath, &'a WorkerCoSQueueFastPath)> {
-    let iface = cos_fast_interface(cos_fast_interfaces, egress_ifindex)?;
-    let queue = iface.queue_fast_path(requested_queue_id)?;
-    Some((iface, queue))
-}
-
-fn redirect_local_cos_request_to_owner(
-    cos_fast_interfaces: &FastMap<i32, WorkerCoSInterfaceFastPath>,
-    req: TxRequest,
-    current_worker_id: u32,
-    worker_commands_by_id: &BTreeMap<u32, Arc<Mutex<VecDeque<WorkerCommand>>>>,
-) -> Result<(), TxRequest> {
-    let Some((iface_fast, queue_fast)) =
-        cos_fast_queue(cos_fast_interfaces, req.egress_ifindex, req.cos_queue_id)
-    else {
-        return Err(req);
-    };
-    if queue_fast.shared_exact && iface_fast.tx_owner_live.is_some() {
-        return Err(req);
-    }
-    let owner_worker_id = queue_fast.owner_worker_id;
-    if owner_worker_id == current_worker_id {
-        return Err(req);
-    }
-    if let Some(owner_live) = queue_fast.owner_live.as_ref() {
-        return owner_live.enqueue_tx_owned(req);
-    }
-    let Some(commands) = worker_commands_by_id.get(&owner_worker_id) else {
-        return Err(req);
-    };
-    if let Ok(mut pending) = commands.lock() {
-        pending.push_back(WorkerCommand::EnqueueShapedLocal(req));
-        return Ok(());
-    }
-    Err(req)
-}
-
-fn redirect_local_cos_request_to_owner_binding(
-    current_live: &Arc<BindingLiveState>,
-    cos_fast_interfaces: &FastMap<i32, WorkerCoSInterfaceFastPath>,
-    req: TxRequest,
-) -> Result<(), TxRequest> {
-    // Caller ordering matters: shared exact queues that already have a local TX
-    // path were filtered out in redirect_local_cos_request_to_owner().
-    let Some(iface_fast) = cos_fast_interface(cos_fast_interfaces, req.egress_ifindex) else {
-        return Err(req);
-    };
-    let Some(owner_live) = iface_fast.tx_owner_live.as_ref() else {
-        return Err(req);
-    };
-    if Arc::ptr_eq(owner_live, current_live) {
-        return Err(req);
-    }
-    owner_live.enqueue_tx_owned(req)
-}
-
-#[inline]
-fn prepared_cos_request_stays_on_current_tx_binding(
-    binding_ifindex: i32,
-    iface_fast: &WorkerCoSInterfaceFastPath,
-    queue_fast: &WorkerCoSQueueFastPath,
-) -> bool {
-    binding_ifindex == iface_fast.tx_ifindex && queue_fast.shared_exact
-}
-
-fn redirect_prepared_cos_request_to_owner(
-    binding: &mut BindingWorker,
-    req: PreparedTxRequest,
-    current_worker_id: u32,
-    worker_commands_by_id: &BTreeMap<u32, Arc<Mutex<VecDeque<WorkerCommand>>>>,
-) -> Result<(), PreparedTxRequest> {
-    let Some((iface_fast, queue_fast)) = cos_fast_queue(
-        &binding.cos_fast_interfaces,
-        req.egress_ifindex,
-        req.cos_queue_id,
-    ) else {
-        return Err(req);
-    };
-    if queue_fast.shared_exact && iface_fast.tx_owner_live.is_some() {
-        return Err(req);
-    }
-    let owner_worker_id = queue_fast.owner_worker_id;
-    if owner_worker_id == current_worker_id {
-        return Err(req);
-    }
-    let Some(frame) = binding
-        .umem
-        .area()
-        .slice(req.offset as usize, req.len as usize)
-        .map(|frame| frame.to_vec())
-    else {
-        return Err(req);
-    };
-    let local_req = TxRequest {
-        bytes: frame,
-        expected_ports: req.expected_ports,
-        expected_addr_family: req.expected_addr_family,
-        expected_protocol: req.expected_protocol,
-        flow_key: req.flow_key.clone(),
-        egress_ifindex: req.egress_ifindex,
-        cos_queue_id: req.cos_queue_id,
-        dscp_rewrite: req.dscp_rewrite,
-    };
-    if redirect_local_cos_request_to_owner(
-        &binding.cos_fast_interfaces,
-        local_req,
-        current_worker_id,
-        worker_commands_by_id,
-    )
-    .is_ok()
-    {
-        recycle_prepared_immediately(binding, &req);
-        return Ok(());
-    }
-    Err(req)
-}
-
-fn redirect_prepared_cos_request_to_owner_binding(
-    binding: &mut BindingWorker,
-    req: PreparedTxRequest,
-) -> Result<(), PreparedTxRequest> {
-    let Some((iface_fast, queue_fast)) = cos_fast_queue(
-        &binding.cos_fast_interfaces,
-        req.egress_ifindex,
-        req.cos_queue_id,
-    ) else {
-        return Err(req);
-    };
-    // Keep shared exact traffic on the current binding when it already sits on
-    // the resolved TX path; redirecting it sideways would force a copy back
-    // into local TX instead of preserving the prepared path.
-    if prepared_cos_request_stays_on_current_tx_binding(binding.ifindex, iface_fast, queue_fast) {
-        return Err(req);
-    }
-    let Some(owner_live) = iface_fast.tx_owner_live.as_ref() else {
-        return Err(req);
-    };
-    if Arc::ptr_eq(owner_live, &binding.live) {
-        return Err(req);
-    }
-    let Some(frame) = binding
-        .umem
-        .area()
-        .slice(req.offset as usize, req.len as usize)
-        .map(|frame| frame.to_vec())
-    else {
-        return Err(req);
-    };
-    let local_req = TxRequest {
-        bytes: frame,
-        expected_ports: req.expected_ports,
-        expected_addr_family: req.expected_addr_family,
-        expected_protocol: req.expected_protocol,
-        flow_key: req.flow_key.clone(),
-        egress_ifindex: req.egress_ifindex,
-        cos_queue_id: req.cos_queue_id,
-        dscp_rewrite: req.dscp_rewrite,
-    };
-    if owner_live.enqueue_tx(local_req).is_ok() {
-        recycle_prepared_immediately(binding, &req);
-        return Ok(());
-    }
-    Err(req)
-}
-
 pub(in crate::afxdp) fn prime_cos_root_for_service(binding: &mut BindingWorker, root_ifindex: i32, now_ns: u64) -> bool {
     let shared_root_lease = binding
         .cos_fast_interfaces
@@ -1475,11 +1226,13 @@ const COS_TIMER_WHEEL_L0_HORIZON_TICKS: u64 = COS_TIMER_WHEEL_L0_SLOTS as u64;
 // non-test builds (Copilot review on PR #976).
 use super::cos::{
     apply_cos_admission_ecn_policy, cos_flow_aware_buffer_limit, cos_flow_bucket_index,
-    cos_item_flow_key,
-    cos_queue_drain_all, cos_queue_flow_share_limit, cos_queue_is_empty, cos_queue_push_back, cos_queue_push_front, cos_queue_restore_front,
-    drain_shaped_tx, ensure_cos_interface_runtime,
-    maybe_top_up_cos_root_lease, park_cos_queue, publish_committed_queue_vtime,
-    release_cos_root_lease, CoSServicePhase,
+    cos_item_flow_key, cos_queue_drain_all, cos_queue_flow_share_limit,
+    cos_queue_is_empty, cos_queue_push_back, cos_queue_push_front, cos_queue_restore_front,
+    drain_shaped_tx, ensure_cos_interface_runtime, maybe_top_up_cos_root_lease, park_cos_queue,
+    prepared_cos_request_stays_on_current_tx_binding, publish_committed_queue_vtime,
+    redirect_local_cos_request_to_owner, redirect_prepared_cos_request_to_owner,
+    redirect_prepared_cos_request_to_owner_binding, release_cos_root_lease,
+    resolve_local_routing_decision, CoSServicePhase, LocalRoutingDecision, Step1Action,
 };
 #[cfg(test)]
 use super::cos::ecn::{ethernet_l3, mark_ecn_ce_ipv4, mark_ecn_ce_ipv6, EthernetL3};
@@ -1487,7 +1240,7 @@ use super::cos::ecn::{ethernet_l3, mark_ecn_ce_ipv4, mark_ecn_ce_ipv6, EthernetL
 use super::cos::{
     account_cos_queue_flow_dequeue, account_cos_queue_flow_enqueue,
     apply_cos_queue_flow_fair_promotion, assign_local_dscp_rewrite, bdp_floor_bytes,
-    build_cos_interface_runtime, cos_batch_tx_made_progress, cos_flow_hash_seed_from_os,
+    build_cos_interface_runtime, cos_batch_tx_made_progress, cos_flow_hash_seed_from_os, redirect_local_cos_request_to_owner_binding,
     cos_guarantee_quantum_bytes, cos_queue_clear_orphan_snapshot_after_drop,
     cos_queue_pop_front, cos_queue_prospective_active_flows,
     cos_queue_v_min_consume_suspension, cos_queue_v_min_continue, count_park_reason,
@@ -2619,7 +2372,7 @@ fn recycle_completed_tx_offset(
     }
 }
 
-pub(super) fn recycle_prepared_immediately(binding: &mut BindingWorker, req: &PreparedTxRequest) {
+pub(in crate::afxdp) fn recycle_prepared_immediately(binding: &mut BindingWorker, req: &PreparedTxRequest) {
     // #760 / Codex review note: when `req.recycle` is
     // `FillOnSlot(fill_slot)` with `fill_slot != binding.slot`,
     // `recycle_cancelled_prepared_offset` routes the frame to THIS
